Converted all* integer types to fixed-with types
* The exception is char, which is still used for strings

+++ src/file-helper.h
@@ -12,12 +12,12 @@ struct FileBin {
 // Displays all content of a file to the screen, using `printf`. This returns
 // `EXIT_SUCCESS` if and only if the entire file was read successfully; 
 // any other return is an error.
-int putFile(char *fileName);
+int32_t putFile(char *fileName);
 
 // Attempts to open a read-only file, displaying a message if the file cannot 
 // be opened. If the file cannot be opened, this returns NULL. The file must be 
 // closed with `fclose()` after usage
-FILE *attemptRFileOpen(char *filename, int maxCount); 
+FILE *attemptRFileOpen(char *filename, uint16_t maxCount); 
 
 // Attempts to create a write-only file, displaying a message if the file cannot
 // be opened. If the file cannot be opened, this returns NULL. If a file already 
@@ -31,14 +31,14 @@ uint64_t fileLength(FILE *file);
 // Puts the entire contents of a read-only file into a byte buffer.
 // This buffer MUST be freed after its used. This returns NULL if any error 
 // occurs
-struct FileBin *readBin(char *filename, int filenameLen);
+struct FileBin *readBin(char *filename, uint16_t filenameLen);
 
 void freeBin(struct FileBin *bin);
 
 // Attempts to open or create a file for writing, displaying a message if the 
 // file cannot be opened. The flags argument specifies whether to override
 // an existing file, to prompt beforehand, or to only ever create new ones.
-FILE *attemptWFileOpen(char *filename, int maxCount, uint32_t flags);
+FILE *attemptWFileOpen(char *filename, uint16_t maxCount, uint32_t flags);
 
 // Attempts to close the given file and then remove it. The file is not removed
 // if either the file could not be closed, the preserve flag is set, or the
